(bmbattle) make bmbattle.h
