+++ data/data_59BADC.s
@@ -1,0 +1,209 @@
+    .section .data
+
+	.global gUnknown_0859BADC
+gUnknown_0859BADC:  @ 0x0859BADC
+	.incbin "baserom.gba", 0x59BADC, 0x40
+
+	.global gUnknown_0859BB1C
+gUnknown_0859BB1C:  @ 0x0859BB1C
+	.incbin "baserom.gba", 0x59BB1C, 0xB8
+
+	.global gUnknown_0859BBD4
+gUnknown_0859BBD4:  @ 0x0859BBD4
+	.incbin "baserom.gba", 0x59BBD4, 0x20
+
+	.global gUnknown_0859BBF4
+gUnknown_0859BBF4:  @ 0x0859BBF4
+	.incbin "baserom.gba", 0x59BBF4, 0x18
+
+	.global gUnknown_0859BC0C
+gUnknown_0859BC0C:  @ 0x0859BC0C
+	.incbin "baserom.gba", 0x59BC0C, 0x18
+
+	.global gUnknown_0859BC24
+gUnknown_0859BC24:  @ 0x0859BC24
+	.incbin "baserom.gba", 0x59BC24, 0x18
+
+	.global gUnknown_0859BC3C
+gUnknown_0859BC3C:  @ 0x0859BC3C
+	.incbin "baserom.gba", 0x59BC3C, 0x28
+
+	.global gUnknown_0859BC64
+gUnknown_0859BC64:  @ 0x0859BC64
+	.incbin "baserom.gba", 0x59BC64, 0x44
+
+	.global gUnknown_0859BCA8
+gUnknown_0859BCA8:  @ 0x0859BCA8
+	.incbin "baserom.gba", 0x59BCA8, 0x4C
+
+	.global gUnknown_0859BCF4
+gUnknown_0859BCF4:  @ 0x0859BCF4
+	.incbin "baserom.gba", 0x59BCF4, 0x4C
+
+	.global gUnknown_0859BD40
+gUnknown_0859BD40:  @ 0x0859BD40
+	.incbin "baserom.gba", 0x59BD40, 0x60
+
+	.global gUnknown_0859BDA0
+gUnknown_0859BDA0:  @ 0x0859BDA0
+	.incbin "baserom.gba", 0x59BDA0, 0x50
+
+	.global gUnknown_0859BDF0
+gUnknown_0859BDF0:  @ 0x0859BDF0
+	.incbin "baserom.gba", 0x59BDF0, 0x20
+
+	.global gUnknown_0859BE10
+gUnknown_0859BE10:  @ 0x0859BE10
+	.incbin "baserom.gba", 0x59BE10, 0x18
+
+	.global gUnknown_0859BE28
+gUnknown_0859BE28:  @ 0x0859BE28
+	.incbin "baserom.gba", 0x59BE28, 0x1188
+
+	.global gUnknown_0859CFB0
+gUnknown_0859CFB0:  @ 0x0859CFB0
+	.incbin "baserom.gba", 0x59CFB0, 0x24
+
+	.global gUnknown_0859CFD4
+gUnknown_0859CFD4:  @ 0x0859CFD4
+	.incbin "baserom.gba", 0x59CFD4, 0x24
+
+	.global gUnknown_0859CFF8
+gUnknown_0859CFF8:  @ 0x0859CFF8
+	.incbin "baserom.gba", 0x59CFF8, 0x24
+
+	.global gUnknown_0859D01C
+gUnknown_0859D01C:  @ 0x0859D01C
+	.incbin "baserom.gba", 0x59D01C, 0x24
+
+	.global gUnknown_0859D040
+gUnknown_0859D040:  @ 0x0859D040
+	.incbin "baserom.gba", 0x59D040, 0x24
+
+	.global gUnknown_0859D064
+gUnknown_0859D064:  @ 0x0859D064
+	.incbin "baserom.gba", 0x59D064, 0x24
+
+	.global gUnknown_0859D088
+gUnknown_0859D088:  @ 0x0859D088
+	.incbin "baserom.gba", 0x59D088, 0x24
+
+	.global gUnknown_0859D0AC
+gUnknown_0859D0AC:  @ 0x0859D0AC
+	.incbin "baserom.gba", 0x59D0AC, 0x24
+
+	.global gUnknown_0859D0D0
+gUnknown_0859D0D0:  @ 0x0859D0D0
+	.incbin "baserom.gba", 0x59D0D0, 0x24
+
+	.global gUnknown_0859D0F4
+gUnknown_0859D0F4:  @ 0x0859D0F4
+	.incbin "baserom.gba", 0x59D0F4, 0x24
+
+	.global gUnknown_0859D118
+gUnknown_0859D118:  @ 0x0859D118
+	.incbin "baserom.gba", 0x59D118, 0x24
+
+	.global gUnknown_0859D13C
+gUnknown_0859D13C:  @ 0x0859D13C
+	.incbin "baserom.gba", 0x59D13C, 0x24
+
+	.global gUnknown_0859D160
+gUnknown_0859D160:  @ 0x0859D160
+	.incbin "baserom.gba", 0x59D160, 0x24
+
+	.global gUnknown_0859D184
+gUnknown_0859D184:  @ 0x0859D184
+	.incbin "baserom.gba", 0x59D184, 0x24
+
+	.global gUnknown_0859D1A8
+gUnknown_0859D1A8:  @ 0x0859D1A8
+	.incbin "baserom.gba", 0x59D1A8, 0x24
+
+	.global gUnknown_0859D1CC
+gUnknown_0859D1CC:  @ 0x0859D1CC
+	.incbin "baserom.gba", 0x59D1CC, 0x24
+
+	.global gUnknown_0859D1F0
+gUnknown_0859D1F0:  @ 0x0859D1F0
+	.incbin "baserom.gba", 0x59D1F0, 0x24
+
+	.global gUnknown_0859D214
+gUnknown_0859D214:  @ 0x0859D214
+	.incbin "baserom.gba", 0x59D214, 0x24
+
+	.global gUnknown_0859D238
+gUnknown_0859D238:  @ 0x0859D238
+	.incbin "baserom.gba", 0x59D238, 0x20
+
+	.global gUnknown_0859D258
+gUnknown_0859D258:  @ 0x0859D258
+	.incbin "baserom.gba", 0x59D258, 0x20
+
+	.global gUnknown_0859D278
+gUnknown_0859D278:  @ 0x0859D278
+	.incbin "baserom.gba", 0x59D278, 0x20
+
+	.global gUnknown_0859D298
+gUnknown_0859D298:  @ 0x0859D298
+	.incbin "baserom.gba", 0x59D298, 0x20
+
+	.global gUnknown_0859D2B8
+gUnknown_0859D2B8:  @ 0x0859D2B8
+	.incbin "baserom.gba", 0x59D2B8, 0x20
+
+	.global gUnknown_0859D2D8
+gUnknown_0859D2D8:  @ 0x0859D2D8
+	.incbin "baserom.gba", 0x59D2D8, 0x20
+
+	.global gUnknown_0859D2F8
+gUnknown_0859D2F8:  @ 0x0859D2F8
+	.incbin "baserom.gba", 0x59D2F8, 0x20
+
+	.global gUnknown_0859D318
+gUnknown_0859D318:  @ 0x0859D318
+	.incbin "baserom.gba", 0x59D318, 0x20
+
+	.global gUnknown_0859D338
+gUnknown_0859D338:  @ 0x0859D338
+	.incbin "baserom.gba", 0x59D338, 0x20
+
+	.global gUnknown_0859D358
+gUnknown_0859D358:  @ 0x0859D358
+	.incbin "baserom.gba", 0x59D358, 0x20
+
+	.global gUnknown_0859D378
+gUnknown_0859D378:  @ 0x0859D378
+	.incbin "baserom.gba", 0x59D378, 0x20
+
+	.global gUnknown_0859D398
+gUnknown_0859D398:  @ 0x0859D398
+	.incbin "baserom.gba", 0x59D398, 0x20
+
+	.global gUnknown_0859D3B8
+gUnknown_0859D3B8:  @ 0x0859D3B8
+	.incbin "baserom.gba", 0x59D3B8, 0x20
+
+	.global gUnknown_0859D3D8
+gUnknown_0859D3D8:  @ 0x0859D3D8
+	.incbin "baserom.gba", 0x59D3D8, 0x20
+
+	.global gUnknown_0859D3F8
+gUnknown_0859D3F8:  @ 0x0859D3F8
+	.incbin "baserom.gba", 0x59D3F8, 0x20
+
+	.global gUnknown_0859D418
+gUnknown_0859D418:  @ 0x0859D418
+	.incbin "baserom.gba", 0x59D418, 0x20
+
+	.global gUnknown_0859D438
+gUnknown_0859D438:  @ 0x0859D438
+	.incbin "baserom.gba", 0x59D438, 0x20
+
+	.global gUnknown_0859D458
+gUnknown_0859D458:  @ 0x0859D458
+	.incbin "baserom.gba", 0x59D458, 0x20
+
+	.global gUnknown_0859D478
+gUnknown_0859D478:  @ 0x0859D478
+	.incbin "baserom.gba", 0x59D478, 0x440
