+++ include/bmunit.h
@@ -318,62 +318,6 @@ enum {
     UNIT_USEBIT_STAFF  = (1 << 1),
 };
 
-// TODO: MOVE ELSEWHERE (bmbattle?)
-struct BattleUnit {
-    /* 00 */ struct Unit unit;
-
-    /* 48 */ u16 weaponAfter;
-    /* 4A */ u16 weaponBefore;
-    /* 4C */ u32 weaponAttributes;
-    /* 50 */ u8 weaponType;
-    /* 51 */ u8 weaponSlotIndex;
-
-    /* 52 */ s8 canCounter;
-
-    /* 53 */ s8 WTHitModifier;
-    /* 54 */ s8 WTAtkModifier;
-
-    /* 55 */ u8 terrainIndex;
-    /* 56 */ s8 terrainDefense;
-    /* 57 */ s8 terrainAvoid;
-    /* 58 */ s8 terrainResistance;
-    /* 59 */ u8 _u59;
-
-    /* 5A */ short battleAttack;
-    /* 5C */ short battleDefense;
-    /* 5E */ short battleAttackSpeed;
-    /* 60 */ short battleHit;
-    /* 62 */ short battleAvoid;
-    /* 64 */ short battleEffectiveHit;
-    /* 66 */ short battleCrit;
-    /* 68 */ short battleDodge;
-    /* 6A */ short battleEffectiveCrit;
-    /* 6C */ short battleSilencerRate;
-
-    /* 6E */ s8 expGain;
-    /* 6F */ s8 statusOut;
-    /* 70 */ u8 levelPrevious;
-    /* 71 */ u8 expPrevious;
-
-    /* 72 */ u8 currentHP;
-
-    /* 73 */ s8 changeHP;
-    /* 74 */ s8 changePow;
-    /* 75 */ s8 changeSkl;
-    /* 76 */ s8 changeSpd;
-    /* 77 */ s8 changeDef;
-    /* 78 */ s8 changeRes;
-    /* 79 */ s8 changeLck;
-    /* 7A */ s8 changeCon;
-
-    /* 7B */ s8 wexpMultiplier;
-    /* 7C */ s8 nonZeroDamage;
-    /* 7D */ s8 weaponBroke;
-
-    /* 7E */ s8 unk7E;
-    /* 7F */ u8 _u7F;
-};
-
 // TODO: MOVE ELSEWHERE
 extern const struct ClassData gClassData[]; // gClassData
 extern const struct CharacterData gCharacterData[]; // gCharacterData
