+++ include/functions.h
@@ -1516,115 +1516,6 @@ s8 CanUnitUseItem(struct Unit* unit, int item);
 // ??? NewTorchStaffSelection(???);
 // ??? CanUnitUseItemPrepScreen(???);
 // ??? sub_802A108(???);
-// ??? sub_802A13C(???);
-// ??? sub_802A20C(???);
-// ??? BATTLE_SaveFromBattle(???);
-// ??? sub_802A318(???);
-// ??? sub_802A350(???);
-// ??? sub_802A364(???);
-// ??? sub_802A384(???);
-// ??? sub_802A398(???);
-// ??? SetupBattleStructFromUnitAndWeapon(???);
-// ??? RollRNIfBattleStarted(???);
-// ??? Roll2RNIfBattleStarted(???);
-void CopyUnitToBattleStruct(struct BattleUnit* bUnit, struct Unit* unit);
-// ??? CopyUnitToBattleStructRawStats(???);
-// ??? WriteBattleStructTerrainBonuses(???);
-// ??? BattleSetupTerrainData(???);
-// ??? SetupBattleWeaponData(???);
-// ??? SetupBattleBallistaWeaponData(???);
-// ??? sub_802A958(???);
-// ??? FillPreBattleStats(???);
-// ??? FillBattleStats(???);
-// ??? BattleApplyMiscBonuses(???);
-// ??? BattleLoadDefense(???);
-// ??? LoadRawDefense(???);
-// ??? BattleLoadAttack(???);
-// ??? BattleLoadAS(???);
-// ??? BattleLoadHit(???);
-// ??? BattleLoadAvoid(???);
-// ??? BattleLoadCrit(???);
-// ??? BattleLoadDodge(???);
-// ??? ComputeHit(???);
-// ??? ComputeCrit(???);
-// ??? ComputeLethalityChance(???);
-// ??? BattleApplySRankBonuses(???);
-// ??? BattleComputeBuffStatus(???);
-// ??? ComputeSpecialWeapons(???);
-// ??? ClearRounds(???);
-// ??? MakeBattle(???);
-// ??? GetBattleUnitPointers(???);
-// ??? BattleCheckDoubling(???);
-// ??? MakeBattleRound(???);
-// ??? GetBattleHitCount(???);
-// ??? BattleCheckBrave(???);
-// ??? CheckForTriangleAttack(???);
-// ??? UpdateBattleStats(???);
-// ??? RollForSureShot(???);
-// ??? RollForPierce(???);
-// ??? RollForGreatShield(???);
-// ??? RollForLethality(???);
-// ??? NullifyBattleDamageIfUsingStone(???);
-// ??? GenerateCurrentRoundData(???);
-// ??? UpdateBattleTriangleAttackData(???);
-// ??? CurrentRound_ComputeWeaponEffects(???);
-// ??? MakeNextBattleHitRound(???);
-// ??? sub_802B92C(???);
-// ??? GetStatIncrease(???);
-int GetAutoleveledStat(int growth, int levelCount);
-// ??? CanUnitNotLevelUp(???);
-void CheckForLevelUp(struct BattleUnit* bUnit);
-// ??? sub_802BC00(???);
-// ??? sub_802BD50(???);
-// ??? sub_802BEA0(???);
-// ??? CheckForLevelUpCaps(???);
-// ??? SaveUnitsFromBattle(???);
-// ??? sub_802C0B0(???);
-// ??? GetBattleNewWExp(???);
-// ??? BattleUnit_DidWRankGoUp(???);
-void SaveUnitFromBattle(struct Unit* unit, struct BattleUnit* bUnit);
-// ??? sub_802C2D4(???);
-// ??? UpdateBallistaUsesFromBattle(???);
-void NullSomeStuff(void);
-// ??? sub_802C344(???);
-// ??? sub_802C368(???);
-// ??? sub_802C398(???);
-// ??? sub_802C3D8(???);
-// ??? sub_802C40C(???);
-// ??? sub_802C450(???);
-// ??? sub_802C4F0(???);
-// ??? sub_802C534(???);
-// ??? HandleSomeExp(???);
-// ??? GetBattleUnitStaffExp(???);
-// ??? InstigatorAdd10Exp(???);
-// ??? sub_802C6EC(???);
-// ??? sub_802C740(???);
-// ??? BattleReverseWTriangeEffect(???);
-// ??? BattleApplyWeaponTriangle(???);
-// ??? DoSomeBattleWeaponStuff(???);
-// ??? MakeSnagBattleTarget(???);
-// ??? FillSnagBattleStats(???);
-// ??? SaveSnagWallFromBattle(???);
-void BeginBattleAnimations(void);
-// ??? sub_802CA70(???);
-// ??? sub_802CA98(???);
-// ??? nullsub_11(???);
-// ??? sub_802CAFC(???);
-// ??? sub_802CB24(???);
-// ??? sub_802CBC8(???);
-// ??? SaveInstigatorFromBattle(???);
-// ??? SaveInstigatorWith10ExtraExp(???);
-// ??? sub_802CC54(???);
-// ??? GetStaffAccuracy(???);
-void sub_802CD64(struct Unit* unit);
-// ??? IsCurrentBattleTriangleAttack(???);
-// ??? DidWeaponBreak(???);
-// ??? sub_802CEBC(???);
-// ??? CurrentRound_ComputeDamage(???);
-// ??? sub_802CF4C(???);
-// ??? sub_802D0BC(???);
-// ??? sub_802D2B4(???);
-// ??? sub_802D2C4(???);
 // ??? sub_802D2E0(???);
 // ??? sub_802D3B8(???);
 // ??? sub_802D3C0(???);
@@ -1707,8 +1598,8 @@ void sub_802E690(void);
 // ??? AddLightRune(???);
 // ??? RemoveLightRune(???);
 // ??? HandleTrapDecay(???);
-// ??? BattleSomethingTrapChangeTerrain(???);
-// ??? NullAllLightRunesTerrain(???);
+void BattleSomethingTrapChangeTerrain(void);
+void NullAllLightRunesTerrain(void);
 struct Trap* GetTrap(unsigned index);
 // ??? ExecStandardHeal(???);
 // ??? ExecRestore(???);
@@ -5941,7 +5832,7 @@ void RegisterChapterTimeAndTurnCount(struct Struct0202BCF0* chData);
 // ??? DoesThisChapterCount(???);
 // ??? GetGameTotalTime(???);
 // ??? GetGameTotalTurnCount2(???);
-// ??? sub_80A44C8(???);
+void sub_80A44C8(struct Unit* unit);
 // ??? sub_80A4534(???);
 // ??? sub_80A4594(???);
 // ??? BWL_AddWinOrLossIdk(???);
@@ -5949,7 +5840,7 @@ void RegisterChapterTimeAndTurnCount(struct Struct0202BCF0* chData);
 // ??? BWL_IncrementStatScreenViews(???);
 // ??? BWL_IncrementDeployCountMaybe(???);
 void BWL_AddTilesMoved(int charId, int amount);
-// ??? BWL_AddExpGained(???);
+void BWL_AddExpGained(int charId, int expGain);
 void StoreSomeUnitSetFlags(int charId);
 // ??? sub_80A48F0(???);
 // ??? sub_80A4904(???);
@@ -5960,7 +5851,7 @@ void StoreSomeUnitSetFlags(int charId);
 // ??? sub_80A49C8(???);
 // ??? sub_80A49FC(???);
 // ??? BWL_AddFavoritismValue(???);
-// ??? sub_80A4AA4(???);
+void sub_80A4AA4(void);
 // ??? sub_80A4B1C(???);
 // ??? sub_80A4B6C(???);
 // ??? sub_80A4B90(???);
